pyproject.toml
[project]
name = "obsidianize"
version = "0.1.0"
description = "Document your project directories as distilled knowledge notes in Obsidian"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12.0",
    "streamlit>=1.35.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.7.0",
    "ollama>=0.2.0",
    "openai>=1.30.0",
    "rich>=13.0.0",
]

[project.scripts]
obsidianize = "obsidianize.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/obsidianize"]
